pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "nad_ch"
version = "0.1.0"
description = "10x National Address Database Collaboration Hub (NAD-CH)"
authors = []

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
flask = "^3.0.1"
boto3 = "^1.34.25"
psycopg2-binary = "^2.9.9"
celery = "^5.3.6"
redis = "^5.0.1"
alembic = "^1.13.1"
botocore = "^1.34.30"
fiona = "^1.9.5"
pandas = "^2.2.0"
geopandas = "^0.14.2"
pyarrow = "^15.0.0"
shapely = "^2.0.2"
flask-login = "^0.6.3"
requests = "^2.31.0"
pyjwt = "^2.8.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
flake8 = "^6.1.0"
coverage = "^7.3.2"
pytest-mock = "^3.12.0"
black = "^23.12.1"
pytest-env = "^1.1.3"

[tool.poetry.scripts]
cli = "nad_ch.main:run_cli"
format = "scripts.format:main"
lint = "flake8.main.cli:main"
start-web = "nad_ch.main:serve_flask_app"
test = "pytest:main"

[tool.pytest_env]
APP_ENV="test"
DATABASE_URL="sqlite:///:memory:"
QUEUE_BROKER_URL="memory://"
QUEUE_BACKEND_URL="cache+memory://"
LOCAL_ALLOWED_DOMAIN = "test.org"

[tool.flake8]
max-line-length = 88

[tool.black]
line-length = 88
include = '\.pyi?$'
